+++ _src/main.js
@@ -1,7 +1,7 @@
 /**
- * @name Nexo
+ * @name Fewu
  * @author Hogger Team, 0xarch(soloev) is the main author.
- * @home_page https://github.com/0xarch/nexo
+ * @home_page https://github.com/0xarch/fewu
  * @demo_page https://0xarch.github.io
  * @license GPL v3.0 and later
  */